+++ src/ProtonVPN.Core/Servers/Models/Server.cs
@@ -35,7 +35,7 @@ namespace ProtonVPN.Core.Servers.Models
         public string Domain { get; }
         public sbyte Status { get; }
         public int Tier { get; }
-        public sbyte Features { get; }
+        public ulong Features { get; }
         public int Load { get; }
         public float Score { get; }
         public LocationResponse LocationResponse { get; }
@@ -51,7 +51,7 @@ namespace ProtonVPN.Core.Servers.Models
             string domain,
             sbyte status,
             int tier,
-            sbyte features,
+            ulong features,
             sbyte load,
             float score,
             LocationResponse locationResponse,
