+++ src/ProtonVPN.Core/Servers/ServerFeatures.cs
@@ -23,33 +23,33 @@ namespace ProtonVPN.Core.Servers
 {
     public class ServerFeatures
     {
-        public const int STANDARD = 0;
-        public const int SECURE_CORE = 1;
-        public const int TOR = 2;
-        public const int P2P = 4;
-        public const int STREAMING = 8;
-        public const int IPV6 = 16;
-        public const int PARTNER = 64;
+        public const ulong STANDARD = 0;
+        public const ulong SECURE_CORE = 1;
+        public const ulong TOR = 2;
+        public const ulong P2P = 4;
+        public const ulong STREAMING = 8;
+        public const ulong IPV6 = 16;
+        public const ulong PARTNER = 64;
 
-        private readonly int _value;
+        private readonly ulong _value;
 
-        public ServerFeatures(int value)
+        public ServerFeatures(ulong value)
         {
             _value = value;
         }
 
-        public static implicit operator int(ServerFeatures item) => item._value;
+        public static implicit operator ulong(ServerFeatures item) => item._value;
 
         public bool IsSecureCore() => IsSecureCore(_value);
         public bool SupportsTor() => SupportsTor(_value);
         public bool SupportsP2P() => SupportsP2P(_value);
 
-        public static bool IsSecureCore(int value) => (value & SECURE_CORE) != 0;
-        public static bool SupportsTor(int value) => (value & TOR) != 0;
-        public static bool SupportsP2P(int value) => (value & P2P) != 0;
-        public static bool SupportsStreaming(int value) => (value & STREAMING) != 0;
-        public static bool SupportsIpV6(int value) => (value & IPV6) != 0;
-        public static bool IsPartner(int value) => (value & PARTNER) != 0;
+        public static bool IsSecureCore(ulong value) => (value & SECURE_CORE) != 0;
+        public static bool SupportsTor(ulong value) => (value & TOR) != 0;
+        public static bool SupportsP2P(ulong value) => (value & P2P) != 0;
+        public static bool SupportsStreaming(ulong value) => (value & STREAMING) != 0;
+        public static bool SupportsIpV6(ulong value) => (value & IPV6) != 0;
+        public static bool IsPartner(ulong value) => (value & PARTNER) != 0;
     }
 
     [Flags]
